f64 insterad of f32, minor docs

+++ spiess_floarian.go
@@ -9,7 +9,7 @@ type SFResult struct {
 }
 
 // ComputeSF computes the Spiess-Florian algorithm
-func ComputeSF(allLinks []*Link, allStops map[string]struct{}, destination string, odMatrix map[string]map[string]float32) *SFResult {
+func ComputeSF(allLinks []*Link, allStops map[string]struct{}, destination string, odMatrix map[string]map[string]float64) *SFResult {
 	// Part 1: Find optimal strategy
 	ops := FindOptimalStrategy(allLinks, allStops, destination)
 	// Part 2: Assign demand according to optimal strategy
